#27
MeasureDate is double. Let's solve it with protocols.
