+++ Sources/WeatherDataProtocol.swift
@@ -34,14 +34,3 @@ public protocol WeatherDataProtocol {
   /// Measured time
   var measuredTime: String  { get }
 }
-
-public extension WeatherDataProtocol {
-
-  /// Measured date
-  public var measuredDate: Date? {
-    let dateFormatter = DateFormatter()
-    dateFormatter.dateFormat = "dd MMM yy HH:mm:ss"
-    
-    return dateFormatter.date(from: measuredTime)
-  }
-}
